+++ src/lib.rs
@@ -1,9 +1,11 @@
 pub mod commands;
 pub mod move_transition;
 pub mod obs;
+pub mod obs_routing;
 pub mod raffle;
 pub mod sdf_effects;
 pub mod stream_fx;
 pub mod themesong;
+pub mod uberduck;
 pub mod user_messages;
 pub mod users;
