+++ src/obs.rs
@@ -1,9 +1,7 @@
 use crate::move_transition;
-use std::time;
-
 use crate::sdf_effects;
 use crate::stream_fx;
-use anyhow::Result;
+use anyhow::{bail, Result};
 use obws;
 use obws::requests::scene_items::{
     Position, Scale, SceneItemTransform, SetTransform,
@@ -13,7 +11,10 @@ use obws::Client as OBSClient;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::thread;
+use std::time;
 use std::time::Duration;
+use subd_types::{Event, UserMessage};
+use tokio::sync::broadcast;
 
 const DEFAULT_SCENE: &str = "Primary";
 const MEME_SCENE: &str = "memes";
@@ -24,10 +25,12 @@ const DEFAULT_SOURCE: &str = "begin";
 const SINGLE_SETTING_VALUE_TYPE: u32 = 0;
 
 const DEFAULT_STREAM_FX_FILTER_NAME: &str = "Default_Stream_FX";
+
 const DEFAULT_SCROLL_FILTER_NAME: &str = "Default_Scroll";
-const DEFAULT_BLUR_FILTER_NAME: &str = "Default_Blur";
 
 const DEFAULT_SDF_EFFECTS_FILTER_NAME: &str = "Default_SDF_Effects";
+
+const DEFAULT_BLUR_FILTER_NAME: &str = "Default_Blur";
 
 const STREAM_FX_INTERNAL_FILTER_NAME: &str = "streamfx-filter-transform";
 const MOVE_VALUE_INTERNAL_FILTER_NAME: &str = "move_value_filter";
@@ -1500,4 +1503,367 @@ async fn move_with_move_source(
     obs_client.filters().set_enabled(filter_enabled).await?;
 
     Ok(())
+}
+
+// I wish we could just define this somewhere else
+pub async fn handle_obs_commands(
+    _tx: &broadcast::Sender<Event>,
+    obs_client: &OBSClient,
+    splitmsg: Vec<String>,
+    msg: UserMessage,
+) -> Result<()> {
+    // This is because Begin doesn't understand Rust
+    let default_source = String::from(DEFAULT_SOURCE);
+
+    // We try and do some parsing on every command here
+    // These may not always be what we want, but they are sensible
+    // defaults used by many commands
+    let source: &str = splitmsg.get(1).unwrap_or(&default_source);
+    // We try and do some parsing on every command here
+    // These may not always be what we want, but they are sensible
+    // defaults used by many commands let source: &str = splitmsg.get(1).unwrap_or(&default_source);
+    let duration: u32 = splitmsg
+        .get(4)
+        .map_or(3000, |x| x.trim().parse().unwrap_or(3000));
+
+    // WE PANICKED!!!!!!!
+    let filter_value = splitmsg
+        .get(3)
+        .map_or(0.0, |x| x.trim().parse().unwrap_or(0.0));
+
+    // NOTE: If we want to extract values like filter_setting_name and filter_value
+    // we need to figure a way to look up the defaults per command
+    // because they could be different types
+    // for now we are going to try and have them be the same
+    // let filter_setting_name = splitmsg.get(2).map_or("", |x| x.as_str());
+    //
+    // the source got fucked up some how
+
+    match splitmsg[0].as_str() {
+        // ================== //
+        // Scrolling Sources //
+        // ================== //
+
+        // !scroll SOURCE SCROLL_SETTING SPEED DURATION (in milliseconds)
+        // !scroll begin x 5 300
+        //
+        // TODO: Stop using handle_user_input
+        "!scroll" => {
+            let default_filter_setting_name = String::from("speed_x");
+
+            // This is ok, because we have a different default
+            let filter_setting_name =
+                splitmsg.get(2).unwrap_or(&default_filter_setting_name);
+
+            let filter_setting_name: String = match filter_setting_name.as_str()
+            {
+                "x" => String::from("speed_x"),
+                "y" => String::from("speed_y"),
+                _ => default_filter_setting_name,
+            };
+
+            // TODO: THIS 2 is SUPERFLUOUS!!!
+            // WE SHOULD RE-WRITE THIS METHOD NOT TO USE IT
+            handle_user_input(
+                source,
+                DEFAULT_SCROLL_FILTER_NAME,
+                &filter_setting_name,
+                filter_value,
+                duration,
+                2,
+                &obs_client,
+            )
+            .await
+        }
+
+        "!blur" => {
+            let filter_value = splitmsg
+                .get(2)
+                .map_or(100.0, |x| x.trim().parse().unwrap_or(100.0));
+
+            // msg.roles.is_twitch_mod()
+            // msg.roles.is_twitch_founder()
+            // msg.roles.is_twitch_staff()
+            // msg.roles.is_twitch_sub()
+            // if msg.roles.is_twitch_vip() {
+
+            // So maybe the source is wrong
+            // maybe the DEFAULT_MOVE_BLUR_FILTER_NAME name is wrong
+            //
+            // the 2 is also problematic
+            // and we aren't pull in the duration
+            update_and_trigger_move_value_filter(
+                source,
+                DEFAULT_BLUR_FILTER_NAME,
+                "Filter.Blur.Size",
+                filter_value,
+                300,
+                // 5000, // duration
+                0,
+                &obs_client,
+            )
+            .await?;
+
+            // }
+
+            Ok(())
+        }
+
+        // Update to take in 2 as a const
+        "!noblur" | "!unblur" => {
+            if msg.roles.is_twitch_mod() {
+                println!("WE GOT A MOD OVER HERE");
+                update_and_trigger_move_value_filter(
+                    source,
+                    DEFAULT_BLUR_FILTER_NAME,
+                    "Filter.Blur.Size",
+                    0.0,
+                    5000,
+                    2,
+                    &obs_client,
+                )
+                .await?;
+            }
+            Ok(())
+        }
+
+        // =============== //
+        // Scaling Sources //
+        // =============== //
+        "!grow" | "!scale" => {
+            let x: f32 = splitmsg
+                .get(2)
+                .and_then(|temp_x| temp_x.trim().parse().ok())
+                .unwrap_or(1.0);
+            let y: f32 = splitmsg
+                .get(3)
+                .and_then(|temp_y| temp_y.trim().parse().ok())
+                .unwrap_or(1.0);
+
+            let base_scale = Scale {
+                x: Some(x),
+                y: Some(y),
+            };
+            trigger_grow(source, &base_scale, x, y, &obs_client).await
+        }
+
+        // ====================== //
+        // 3D Transforming Sources//
+        // ====================== //
+
+        // This shit is annoying
+        // I almost want to divide it into 3 commands
+        // based on Camera Type
+        // and we have all 3
+        // that might be too much
+        // but i also might be exactly what we want
+        // only spin is wonky
+        // Should also add !spinz
+        "!spin" | "!spinx" | "spiny" => {
+            // HMMMMM
+            let default_filter_setting_name = String::from("z");
+            let filter_setting_name =
+                splitmsg.get(2).unwrap_or(&default_filter_setting_name);
+
+            spin(
+                source,
+                filter_setting_name,
+                filter_value,
+                duration,
+                &obs_client,
+            )
+            .await
+        }
+
+        "!hide" => hide_sources(MEME_SCENE, &obs_client).await,
+        "!show" => set_enabled(MEME_SCENE, source, true, &obs_client).await,
+        "!def_ortho" => default_ortho(source, duration, &obs_client).await,
+        "!ortho" => {
+            if splitmsg.len() < 3 {
+                return Ok(());
+            };
+
+            let filter_setting_name = &splitmsg[2];
+
+            trigger_ortho(
+                source,
+                "3D_Orthographic",
+                filter_setting_name,
+                filter_value,
+                duration,
+                &obs_client,
+            )
+            .await
+        }
+
+        "!perp" => {
+            if splitmsg.len() < 3 {
+                return Ok(());
+            };
+
+            let filter_setting_name = &splitmsg[2];
+
+            trigger_ortho(
+                source,
+                "3D_Perspective",
+                filter_setting_name,
+                filter_value,
+                duration,
+                &obs_client,
+            )
+            .await
+        }
+
+        "!corner" => {
+            if splitmsg.len() < 3 {
+                return Ok(());
+            };
+
+            let filter_setting_name = &splitmsg[2];
+
+            trigger_ortho(
+                source,
+                "3D_CornerPin",
+                filter_setting_name,
+                filter_value,
+                duration,
+                &obs_client,
+            )
+            .await
+        }
+        // Perspective
+        // Corner Pin
+        // Orthographic
+
+        // !3d SOURCE FILTER_NAME FILTER_VALUE DURATION
+        // !3d begin Rotation.Z 3600 5000
+        //
+        // TODO: This is NOT Working!
+        "!3d" => {
+            // If we don't at least have a filter_name, we can't proceed
+            if splitmsg.len() < 3 {
+                bail!("We don't have a filter name, can't proceed");
+            }
+
+            let filter_setting_name = &splitmsg[2];
+
+            trigger_3d(
+                source,
+                filter_setting_name,
+                filter_value,
+                duration,
+                &obs_client,
+            )
+            .await
+        }
+
+        // ============== //
+        // Moving Sources //
+        // ============== //
+        "!move" => {
+            // TODO: Look at this fanciness
+            //       cafce25: if let [source, x, y, ..] = splitmsg {...}
+            if splitmsg.len() > 3 {
+                let source = splitmsg[1].as_str();
+                let x: f32 = splitmsg[2].trim().parse().unwrap_or(0.0);
+                let y: f32 = splitmsg[3].trim().parse().unwrap_or(0.0);
+
+                move_source(source, x, y, &obs_client).await
+            } else {
+                Ok(())
+            }
+        }
+
+        // TODO: I'd like one-for every corner
+        "!tr" => top_right(source, &obs_client).await,
+
+        "!bl" => bottom_right(source, &obs_client).await,
+
+        // ================ //
+        // Compound Effects //
+        // ================ //
+        "!norm" => norm(&source, &obs_client).await,
+
+        "!follow" => {
+            let scene = DEFAULT_SCENE;
+            let leader = splitmsg.get(1).unwrap_or(&default_source);
+            let source = leader;
+
+            follow(source, scene, leader, &obs_client).await
+        }
+        "!staff" => staff(DEFAULT_SOURCE, &obs_client).await,
+
+        // =============================== //
+        // Create Scenes, Sources, Filters //
+        // =============================== //
+        "!create_source" => {
+            let new_scene: obws::requests::scene_items::CreateSceneItem =
+                obws::requests::scene_items::CreateSceneItem {
+                    scene: DEFAULT_SCENE,
+                    source: &source,
+                    enabled: Some(true),
+                };
+
+            // TODO: Why is this crashing???
+            obs_client.scene_items().create(new_scene).await?;
+
+            Ok(())
+        }
+
+        // TEMP: This is for temporary testing!!!!
+        "!split" => {
+            create_split_3d_transform_filters(source, &obs_client).await
+        }
+
+        // This sets up OBS for Begin's current setup
+        "!create_filters_for_source" => {
+            create_filters_for_source(source, &obs_client).await
+        }
+
+        // ========================== //
+        // Show Info About OBS Setup  //
+        // ========================== //
+        // "!filter" => {
+        //     let (_command, words) = msg.message_text.split_once(" ").unwrap();
+
+        //     // TODO: Handle this error
+        //     let details =
+        //         print_filter_info(&source, words, &obs_client)
+        //             .await?;
+        //     client
+        //         .say(twitch_username.clone(), format!("{:?}", details))
+        //         .await
+        // }
+
+        // TODO: Take in Scene
+        "!source" => {
+            print_source_info(source, DEFAULT_SCENE, &obs_client).await
+        }
+
+        "!outline" => {
+            let source = splitmsg[1].as_str();
+            outline(source, &obs_client).await
+        }
+
+        // ====================== //
+        // Show / Hide Subscenes //
+        // ====================== //
+        "!memes" => {
+            set_enabled(DEFAULT_SCENE, MEME_SCENE, true, &obs_client).await
+        }
+
+        "!nomemes" | "!nojokes" | "!work" => {
+            set_enabled(DEFAULT_SCENE, MEME_SCENE, false, &obs_client).await
+        }
+
+        // ==================== //
+        // Change Scenes in OBS //
+        // ==================== //
+        // Rename These Commands
+        "!chat" => trigger_hotkey("OBS_KEY_L", &obs_client).await,
+
+        "!code" => trigger_hotkey("OBS_KEY_H", &obs_client).await,
+
+        _ => Ok(()),
+    }
 }
